Se completa ms7, se deja pendiente ms6

+++ ms5/src/main/java/com/tingeso/ms5/Entity/ReserveEntity.java
@@ -31,7 +31,7 @@ public class ReserveEntity {
     private UserEntity cliente;
 
     private int cantidadPersonas;
-    private int vueltasOTiempo;
+    private double vueltasOTiempo;
     private double montoFinal;
 
     @OneToMany(mappedBy = "reserve", cascade = CascadeType.ALL, orphanRemoval = true)
